move pact tests to a dedicated directory
